# api/openapi.json
{
  "components": {
    "responses": {
      "Internal": {
        "content": {
          "application/problem+msgpack": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "`internal`: unexpected server error."
      },
      "InvalidBody": {
        "content": {
          "application/problem+msgpack": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "`invalid_body`: the body failed decoding or validation."
      },
      "InvalidParameter": {
        "content": {
          "application/problem+msgpack": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "`invalid_parameter`: a query or header parameter is invalid."
      },
      "NotAcceptable": {
        "content": {
          "application/problem+msgpack": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "`not_acceptable`: `Accept` excludes application/vnd.msgpack."
      },
      "NotFound": {
        "content": {
          "application/problem+msgpack": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "`not_found`: no such resource in the caller's scope."
      },
      "PayloadTooLarge": {
        "content": {
          "application/problem+msgpack": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "`payload_too_large`: the body exceeds the route's limit."
      },
      "Restricted": {
        "content": {
          "application/problem+msgpack": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "`account_deletion_pending`: the account is scheduled for deletion and this session is export-only; `password_change_required`: the password was reset and must be changed with `PATCH /me` first."
      },
      "Unauthorized": {
        "content": {
          "application/problem+msgpack": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "`unauthorized`: missing, invalid, expired or revoked access token (signed out, device removed, account disabled)."
      },
      "UnsupportedMediaType": {
        "content": {
          "application/problem+msgpack": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "`unsupported_media_type`: the body is not application/vnd.msgpack."
      }
    },
    "schemas": {
...
      "CreateTaskRequest": {
        "description": "`POST /tasks`.",
        "properties": {
          "due": {
            "description": "📅",
            "format": "date",
            "type": [
              "string",
              "null"
            ]
          },
          "force": {
            "description": "Create even if it looks like a duplicate (records keep-both).",
            "type": "boolean"
          },
          "id": {
            "description": "Client-generated block ID (`t-<ulid>`).",
            "type": [
              "string",
              "null"
            ]
          },
          "note_id": {
            "description": "Home note (default `tasks/Tasks.md`, under a heading per month).",
            "format": "ulid",
            "type": [
              "string",
              "null"
            ]
          },
          "priority": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/TaskPriority",
                "description": "Priority."
              },
              {
                "type": "null"
              }
            ]
          },
          "recurrence": {
            "description": "🔁 phrase in the Tasks plugin's language (`every month on the 1st`).",
            "type": [
              "string",
              "null"
            ]
          },
          "reminders": {
            "description": "Reminders.",
            "items": {
              "$ref": "#/components/schemas/ReminderAt"
            },
            "type": "array"
          },
          "scheduled": {
            "description": "⏳",
            "format": "date",
            "type": [
              "string",
              "null"
            ]
          },
          "start": {
            "description": "🛫",
            "format": "date",
            "type": [
              "string",
              "null"
            ]
          },
          "text": {
            "description": "Description (may contain `[[links]]` and `#tags`).",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "type": "object"
      },
...
      "TaskList": {
        "description": "Tasks.",
        "properties": {
          "items": {
            "description": "In the view's order.",
            "items": {
              "$ref": "#/components/schemas/Task"
            },
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "TaskPriority": {
        "description": "Priority (Tasks plugin signifiers 🔺⏫🔼🔽⏬; absent = normal).",
        "enum": [
          "highest",
          "high",
          "medium",
          "low",
          "lowest"
        ],
        "type": "string"
      },
...
      "TaskViewKind": {
        "description": "List view.",
        "enum": [
          "today",
          "upcoming",
          "overdue",
          "recurring",
          "done",
          "all"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "description": "Short-lived access token (PLAN §8).",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Strata HTTP API. Every body is MessagePack (application/vnd.msgpack); errors are RFC 7807 problem details (application/problem+msgpack). See docs/WIRE_FORMAT.md.",
    "title": "Strata API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tasks": {
      "get": {
        "operationId": "list_tasks",
        "parameters": [
          {
            "description": "View (default `all`); \"today\" is the user's time zone's.",
            "in": "query",
            "name": "view",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TaskViewKind"
            }
          },
          {
            "description": "Only tasks linking to this entity.",
            "in": "query",
            "name": "entity",
            "required": false,
            "schema": {
              "format": "ulid",
              "type": "string"
            }
          },
          {
            "description": "Only tasks in this note.",
            "in": "query",
            "name": "note",
            "required": false,
            "schema": {
              "format": "ulid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/TaskList"
                }
              }
            },
            "description": "Tasks."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Restricted"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "422": {
            "$ref": "#/components/responses/InvalidParameter"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "summary": "Tasks of a view.",
        "tags": [
          "tasks"
        ]
      },
      "post": {
        "operationId": "create_task",
        "requestBody": {
          "content": {
            "application/vnd.msgpack": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/vnd.msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            },
            "description": "Created; one `user: task create <path>` commit."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Restricted"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "409": {
            "content": {
              "application/problem+msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`duplicate_candidates`."
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/InvalidBody"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "summary": "Create a task line (duplicate check unless `force`).",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/{id}": {
      "patch": {
        "operationId": "patch_task",
        "parameters": [
          {
            "description": "Task block ID (`t-<ulid>`).",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional: the task's current version.",
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/vnd.msgpack": {
              "schema": {
                "$ref": "#/components/schemas/PatchTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/vnd.msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            },
            "description": "Updated."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Restricted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "409": {
            "content": {
              "application/problem+msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`version_conflict`."
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/InvalidBody"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "summary": "Edit a task line in place (`If-Match`: the task's version).",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/{id}/cancel": {
      "post": {
        "operationId": "cancel_task",
        "parameters": [
          {
            "description": "Task block ID.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional: the task's current version.",
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTransitioned"
                }
              }
            },
            "description": "Cancelled."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Restricted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "409": {
            "content": {
              "application/problem+msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`task_state_conflict` or `version_conflict`."
          },
          "422": {
            "$ref": "#/components/responses/InvalidParameter"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "summary": "Cancel a task.",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/{id}/complete": {
      "post": {
        "operationId": "complete_task",
        "parameters": [
          {
            "description": "Task block ID.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional: the task's current version.",
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTransitioned"
                }
              }
            },
            "description": "Completed."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Restricted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "409": {
            "content": {
              "application/problem+msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`task_state_conflict` or `version_conflict`."
          },
          "422": {
            "$ref": "#/components/responses/InvalidParameter"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "summary": "Complete a task (a recurring task gets its next occurrence above the done line).",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/{id}/reopen": {
      "post": {
        "operationId": "reopen_task",
        "parameters": [
          {
            "description": "Task block ID.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional: the task's current version.",
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTransitioned"
                }
              }
            },
            "description": "Reopened."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Restricted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "409": {
            "content": {
              "application/problem+msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`task_state_conflict` or `version_conflict`."
          },
          "422": {
            "$ref": "#/components/responses/InvalidParameter"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "summary": "Reopen a done or cancelled task.",
        "tags": [
          "tasks"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}